% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_warrior.R
\name{play_beginner}
\alias{play_beginner}
\title{Play R Warrior}
\usage{
play_beginner(
  ai,
  level = 1,
  warrior_name = "Fisher",
  sleep = getOption("Rwarrior.sleep", 0.6)
)
}
\arguments{
\item{ai}{AI function to control your warrior}

\item{level}{Level number.}

\item{warrior_name}{Name of your warrior, for flavour.}

\item{sleep}{Time between text updates. Set to "prompt" to only progress when pressing the return key.}
}
\value{
A logical that is TRUE on successfully getting to the stairs
}
\description{
Play R Warrior
}
\examples{
AI <- AI <- function(warrior, memory) {
  warrior$walk()
}
play_beginner(AI, level = 1)
}
