% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_warrior.R
\name{play_warrior}
\alias{play_warrior}
\title{Play R Warrior}
\usage{
play_warrior(ai, level = 1, warrior_name = "Fisher", sleep = 0.5)
}
\arguments{
\item{ai}{AI function to control your warrior}

\item{level}{Level number.}

\item{warrior_name}{Name of your warrior, for flavour.}

\item{sleep}{Time between text updates.}
}
\value{
A logical that is TRUE on successfully getting to the exit
}
\description{
Play R Warrior
}
\examples{
AI <- AI <- function(warrior, memory) {
  warrior$walk()
}
play_warrior(AI, level = 1)
}
