% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_warrior.R
\name{play_warrior_inbuilt_levels}
\alias{play_warrior_inbuilt_levels}
\title{For inbuilt levels}
\usage{
play_warrior_inbuilt_levels(
  ai,
  level = 1,
  warrior_name = "Fisher",
  tower = "beginner",
  sleep = 0,
  practice = FALSE,
  debug = FALSE,
  output = FALSE,
  max_turns = 100L
)
}
\description{
For inbuilt levels
}
