% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_readme.R
\name{level_readme}
\alias{level_readme}
\title{Level read me}
\usage{
level_readme(level = 1)
}
\arguments{
\item{level}{The level number (or custom level).}
}
\description{
Level read me
}
\examples{

level_readme(1)
}
